+++ database/migrations/2025_12_29_023937_create_mystery_box_history_table.php
@@ -12,17 +12,19 @@ return new class() extends Migration
      */
     public function up(): void
     {
-        Schema::create('mystery_box_history', function (Blueprint $table) {
-            $table->integer('id_gacha_history')->autoIncrement();
+        if (! Schema::hasTable('mystery_box_history')) {
+            Schema::create('mystery_box_history', function (Blueprint $table) {
+                $table->integer('id_gacha_history')->autoIncrement();
 
-            $table->integer('id_user');
-            $table->integer('id_product');
+                $table->integer('id_user');
+                $table->integer('id_product');
 
-            $table->timestamps();
+                $table->timestamps();
 
-            $table->foreign('id_user')->references('id_user')->on('user')->onDelete('cascade');
-            $table->foreign('id_product')->references('id_product')->on('product')->onDelete('cascade');
-        });
+                $table->foreign('id_user')->references('id_user')->on('user')->onDelete('cascade');
+                $table->foreign('id_product')->references('id_product')->on('product')->onDelete('cascade');
+            });
+        }
     }
 
     /**
